Flattened out the class structure. Moved all nested classes to individual class files.
