+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Choice.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    class Choice
+    {
+        public string text;
+        public int index;
+        public string finish_reason;
+    }
+}
