+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Choice.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09460da41ac90854e807aa7f6a0dc57c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
