+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/DefinitionData.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    public class DefinitionData
+    {
+        public object[][] sseq;
+    }
+}
