+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/DefinitionData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 034ccbbe015a1904f9e56f86e52901ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
