+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/HwiData.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    public class HwiData
+    {
+        public string hw;
+        public PrsData[] prs;
+    }
+}
