+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/HwiData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07306036df73a94458a2d2bcbb9e6a14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
