+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Item.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    class Item
+    {
+        public MetaData meta;
+        public HwiData hwi;
+        public string fl;
+        public DefinitionData[] def;
+        public string[] shortdef;
+    }
+}
