+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Item.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 434ef3eaf0590dd47aa57cad43cae5d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
