+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/MetaData.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    public class MetaData
+    {
+        public string id;
+        public string uuid;
+        public string sort;
+        public string src;
+        public string section;
+        public string[] stems;
+        public bool offensive;
+    }
+}
