+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/MetaData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 337c7f1ac46b6814cabe733685e3c3e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
