+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/PromptRequest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ae4d06c776ce404f8c0e85b44341e94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
