+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/PrsData.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    public class PrsData
+    {
+        public string mw;
+        public SoundData sound;
+    }
+}
