+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/PrsData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abd74338398876c4481cc6d1754ab206
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
