+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Response.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0d808d2b65dd094f93ba674cb5b5a1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
