+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/SoundData.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Description
+{
+    [Serializable]
+    public class SoundData
+    {
+        public string audio;
+    }
+}
