+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/SoundData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19342e3f51da88b4095f60501f145da3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
