+++ Lorem Ipsum XR Translator/Assets/Scripts/Description/Usage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a82b62b289049d3419118c6433407ec4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
