+++ Lorem Ipsum XR Translator/ProjectSettings/TagManager.asset
@@ -36,7 +36,7 @@ TagManager:
   - 
   - 
   - 
-  - 
+  - Spatial Awareness
   m_SortingLayers:
   - name: Default
     uniqueID: 0
